Rename filenames mesh.h-cpp in mesh_set.h-cpp
