+++ tests/mesh/test-connected-components.cpp
@@ -37,9 +37,9 @@
 
 #include <vector>
 
-#include <ringmesh/mesh/mesh.h>
 #include <ringmesh/mesh/mesh_builder.h>
 #include <ringmesh/mesh/mesh_index.h>
+#include <ringmesh/mesh/mesh_set.h>
 
 /*!
  * @author Benjamin Chauvin
